+++ ckanext/dcatde/tests/test_dcatde.py
@@ -10,7 +10,7 @@ import pprint
 import pkg_resources
 
 from rdflib import Graph, URIRef, Literal
-from rdflib.namespace import Namespace
+from rdflib.namespace import Namespace, RDF
 
 from ckanext.dcat.profiles import EuropeanDCATAPProfile
 from ckanext.dcat.processors import RDFParser
@@ -35,7 +35,7 @@ class TestDCATde(unittest.TestCase):
     SPDX = Namespace('http://spdx.org/rdf/terms#')
 
     # own namespace
-    DCATDE = Namespace("http://dcat-ap.de/def/dcatde/1_0/")
+    DCATDE = Namespace("http://dcat-ap.de/def/dcatde/1.0.1/")
 
     dcat_theme_prefix = "http://publications.europa.eu/resource/authority/data-theme/"
 
@@ -59,10 +59,18 @@ class TestDCATde(unittest.TestCase):
 
     predicate_pattern = re.compile("[a-zA-Z]:[a-zA-Z]")
 
+    def addLanguages(self, rdf_parser, dataset_ref, subject, predicate, text):
+        object_refs = [d for d in rdf_parser.g.objects(dataset_ref, subject)]
+        self.assertEqual(len(object_refs), 1)
+        object_ref = object_refs[0]
+        rdf_parser.g.add((object_ref, predicate, Literal(text + u' (DE)', lang='de')))
+        rdf_parser.g.add((object_ref, predicate, Literal(text + u' (EN)', lang='en')))
+
     def _add_basic_fields_with_languages(self, rdf_parser):
         dataset_refs = [d for d in rdf_parser._datasets()]
         self.assertEqual(len(dataset_refs), 1)
         dataset_ref = dataset_refs[0]
+        # Dataset
         rdf_parser.g.add((dataset_ref, self.DCT.title, Literal(u'Naturräume Geest und Marsch (DE)', lang='de')))
         rdf_parser.g.add((dataset_ref, self.DCT.title, Literal(u'Naturräume Geest und Marsch (EN)', lang='en')))
         rdf_parser.g.add((dataset_ref,
@@ -71,7 +79,11 @@ class TestDCATde(unittest.TestCase):
         rdf_parser.g.add((dataset_ref,
                           self.DCT.description,
                           Literal(u'Die Zuordnung des Hamburger Stadtgebietes zu den Naturräumen Geest und Marsch wird dargestellt. (EN)', lang='en')))
-
+        # Publisher
+        self.addLanguages(rdf_parser, dataset_ref, self.DCT.publisher, self.FOAF.name, u'Behörde für Umwelt und Energie (BUE), Amt für Umweltschutz')
+        # ContactPoint
+        self.addLanguages(rdf_parser, dataset_ref, self.DCAT.contactPoint, self.VCARD.fn, u'Herr Dr. Michael Schröder')
+        # Distributions
         distribution_refs = [d for d in rdf_parser.g.objects(dataset_ref, self.DCAT.distribution)]
         self.assertEqual(len(distribution_refs), 2)
         for dist_ref in distribution_refs:
@@ -191,9 +203,9 @@ class TestDCATde(unittest.TestCase):
                 res.get('description'),
                 u'Das ist eine deutsche Beschreibung der Distribution %s (%s)' % (number, lang_string))
 
-    def _get_max_rdf(self):
+    def _get_max_rdf(self, item_name="metadata_max"):
         data = pkg_resources.resource_string(__name__,
-                                             "resources/metadata_max.rdf")
+                                             "resources/%s.rdf" % item_name)
 
         return data
 
@@ -256,8 +268,8 @@ class TestDCATde(unittest.TestCase):
                 "provenance": "dct:provenance",
                 "politicalGeocodingLevelURI": "dcatde:politicalGeocodingLevelURI",
                 "politicalGeocodingURI": ["dcatde:politicalGeocodingURI"],
-                "geocodingText": ["dcatde:geocodingText"],
-                "legalbasisText": ["dcatde:legalbasisText"],
+                "geocodingText": ["dcatde:geocodingDescription"],
+                "legalbasisText": ["dcatde:legalBasis"],
 
                 "temporal_start": "2017-07-06T13:08:40",
                 "temporal_end": "2017-07-06T13:08:41",
@@ -301,6 +313,57 @@ class TestDCATde(unittest.TestCase):
                 })
             }]
         }
+
+    def _build_and_parse_format_mediatype_graph(self, format_item=None, mediatype_item=None):
+        g = Graph()
+
+        dataset = URIRef("http://example.org/datasets/1")
+        g.add((dataset, RDF.type, self.DCAT.Dataset))
+
+        distribution = URIRef("http://example.org/datasets/1/ds/1")
+        g.add((dataset, self.DCAT.distribution, distribution))
+        g.add((distribution, RDF.type, self.DCAT.Distribution))
+        if format_item:
+            g.add((distribution, self.DCT['format'], format_item))
+        if mediatype_item:
+            g.add((distribution, self.DCAT.mediaType, mediatype_item))
+        if format_item is None and mediatype_item is None:
+            raise AssertionError('At least one of format or mediaType is required!')
+
+        p = RDFParser(profiles=['euro_dcat_ap', 'dcatap_de'])
+
+        p.g = g
+
+        dataset = [d for d in p.datasets()][0]
+        return dataset.get('resources')
+
+    def _build_graph_and_check_format_mediatype(self, dataset_dict, expected_format, expected_mediatype):
+        """
+        Creates a graph based on the given dict and checks for dct:format and dct:mediaType in the
+        first resource element.
+
+        :param dataset_dict:
+            dataset dict, expected to contain one resource
+        :param expected_format:
+            expected list of dct:format items in the resource
+        :param expected_mediatype:
+            expected list of dcat:mediaType items in the resource
+        """
+        self.graph = rdflib.Graph()
+        dataset_ref = URIRef("http://example.org/datasets/1")
+
+        dcat = EuropeanDCATAPProfile(self.graph, False)
+        dcat.graph_from_dataset(dataset_dict, dataset_ref)
+
+        dcatde = DCATdeProfile(self.graph, False)
+        dcatde.graph_from_dataset(dataset_dict, dataset_ref)
+
+        # graph should contain the expected nodes
+        resource_ref = list(self.graph.objects(dataset_ref, self.DCAT.distribution))[0]
+        dct_format = list(self.graph.objects(resource_ref, self.DCT['format']))
+        dcat_mediatype = list(self.graph.objects(resource_ref, self.DCAT.mediaType))
+        self.assertEqual(expected_format, dct_format)
+        self.assertEqual(expected_mediatype, dcat_mediatype)
 
     def test_graph_from_dataset(self):
         """ test dcat and dcatde profiles """
@@ -380,9 +443,9 @@ class TestDCATde(unittest.TestCase):
                          self._get_value_from_extras(extras, "is_version_of"))
         self._assert_list(dataset_ref, self.DCATDE.politicalGeocodingURI,
                          self._get_value_from_extras(extras, "politicalGeocodingURI"))
-        self._assert_list(dataset_ref, self.DCATDE.geocodingText,
+        self._assert_list(dataset_ref, self.DCATDE.geocodingDescription,
                          self._get_value_from_extras(extras, "geocodingText"))
-        self._assert_list(dataset_ref, self.DCATDE.legalbasisText,
+        self._assert_list(dataset_ref, self.DCATDE.legalBasis,
                          self._get_value_from_extras(extras, "legalbasisText"))
         self._assert_list(dataset_ref, self.DCATDE.contributorID,
                          self._get_value_from_extras(extras, "contributorID"))
@@ -435,8 +498,88 @@ class TestDCATde(unittest.TestCase):
         # contactPoint
         self._assert_contact_point(dataset_ref, [self.VCARD.fn, self.VCARD.hasEmail])
 
-    def test_parse_dataset(self):
-        maxrdf = self._get_max_rdf()
+    def test_graph_from_dataset_format_iana_uri(self):
+        dataset_dict = self._get_default_dataset_dict()
+        # when only format is available and it looks like an IANA media type, use DCAT.mediaType instead
+        # of DCT.format for output
+        fmt_uri = 'https://www.iana.org/assignments/media-types/application/json'
+        dataset_dict['resources'][0]['format'] = fmt_uri
+        dataset_dict['resources'][0].pop('mimetype')
+
+        # expect no dct:format node and the URI in dcat:mediaType
+        self._build_graph_and_check_format_mediatype(
+            dataset_dict,
+            [],
+            [URIRef(fmt_uri)]
+        )
+
+    def test_graph_from_dataset_format_other_uri(self):
+        dataset_dict = self._get_default_dataset_dict()
+        # when only format is available and it does not look like an IANA media type, use dct:format
+        fmt_uri = 'https://example.com/my/format'
+        dataset_dict['resources'][0]['format'] = fmt_uri
+        dataset_dict['resources'][0].pop('mimetype')
+        dataset_ref = URIRef("http://example.org/datasets/1")
+
+        # expect dct:format node with the URI and no dcat:mediaType
+        self._build_graph_and_check_format_mediatype(
+            dataset_dict,
+            [URIRef(fmt_uri)],
+            []
+        )
+
+    def test_graph_from_dataset_format_mediatype_text(self):
+        dataset_dict = self._get_default_dataset_dict()
+        # if format value looks like an IANA media type, output dcat:mediaType instead of dct:format
+        fmt_text = 'application/json'
+        dataset_dict['resources'][0]['format'] = fmt_text
+        dataset_dict['resources'][0].pop('mimetype')
+
+        # expect no dct:format node and the literal value in dcat:mediaType
+        self._build_graph_and_check_format_mediatype(
+            dataset_dict,
+            [],
+            [Literal(fmt_text)]
+        )
+
+    def test_graph_from_dataset_format_mediatype_same(self):
+        dataset_dict = self._get_default_dataset_dict()
+        # if format and mediaType are identical, output only dcat:mediaType
+        fmt_text = 'application/json'
+        dataset_dict['resources'][0]['format'] = fmt_text
+        dataset_dict['resources'][0]['mimetype'] = fmt_text
+
+        # expect no dct:format node and the literal value in dcat:mediaType
+        self._build_graph_and_check_format_mediatype(
+            dataset_dict,
+            [],
+            [Literal(fmt_text)]
+        )
+
+    def test_graph_from_dataset_format_mediatype_different(self):
+        dataset_dict = self._get_default_dataset_dict()
+        # if format and mediaType are different, output both
+        dataset_dict['resources'][0]['format'] = 'myformat'
+        dataset_dict['resources'][0]['mimetype'] = 'application/json'
+
+        # expect both nodes
+        self._build_graph_and_check_format_mediatype(
+            dataset_dict,
+            [Literal('myformat')],
+            [Literal('application/json')]
+        )
+
+    def test_parse_dataset_v1_0_1(self):
+        self._run_parse_dataset('metadata_max')
+
+    def test_parse_dataset_v1_0(self):
+        self._run_parse_dataset('metadata_max_1_0')
+
+    def test_parse_dataset_multi_namespaces(self):
+        self._run_parse_dataset('metadata_max_multi_namespaces')
+
+    def _run_parse_dataset(self, max_rdf_file):
+        maxrdf = self._get_max_rdf(max_rdf_file)
 
         p = RDFParser(profiles=['euro_dcat_ap', 'dcatap_de'])
 
@@ -485,11 +628,11 @@ class TestDCATde(unittest.TestCase):
         self._assert_extras_string(extras, 'politicalGeocodingLevelURI',
                                   'http://dcat-ap.de/def/politicalGeocoding/Level/state')
 
-        # dcatde:legalbasisText
+        # dcatde:legalBasis
         self._assert_extras_list_serialized(extras, 'legalbasisText',
                                            ['Umweltinformationsgesetz (UIG)'])
 
-        # dcatde:geocodingText
+        # dcatde:geocodingDescription
         self._assert_extras_list_serialized(extras, 'geocodingText',
                                            ['Hamburg'])
 
@@ -568,7 +711,12 @@ class TestDCATde(unittest.TestCase):
         self.assertEqual(
             dataset.get('notes'),
             u'Die Zuordnung des Hamburger Stadtgebietes zu den Naturräumen Geest und Marsch wird dargestellt. (EN)')
-
+        # Publisher and ContactPoint
+        extras = dataset.get('extras')
+        self.assertTrue(len(extras) > 0)
+        self._assert_extras_string(extras, 'publisher_name', u'Behörde für Umwelt und Energie (BUE), Amt für Umweltschutz (EN)')
+        self._assert_extras_string(extras, 'contact_name', u'Herr Dr. Michael Schröder (EN)')
+        # Resources
         self._assert_resource_lang(dataset, 'EN')
 
     @helpers.change_config('ckan.locale_default', 'de')
@@ -589,7 +737,12 @@ class TestDCATde(unittest.TestCase):
         self.assertEqual(
             dataset.get('notes'),
             u'Die Zuordnung des Hamburger Stadtgebietes zu den Naturräumen Geest und Marsch wird dargestellt. (DE)')
-
+        # Publisher and ContactPoint
+        extras = dataset.get('extras')
+        self.assertTrue(len(extras) > 0)
+        self._assert_extras_string(extras, 'publisher_name', u'Behörde für Umwelt und Energie (BUE), Amt für Umweltschutz (DE)')
+        self._assert_extras_string(extras, 'contact_name', u'Herr Dr. Michael Schröder (DE)')
+        # Resources
         self._assert_resource_lang(dataset, 'DE')
 
     @helpers.change_config('ckan.locale_default', 'fr')
@@ -610,7 +763,12 @@ class TestDCATde(unittest.TestCase):
         self.assertIn(
             u'Die Zuordnung des Hamburger Stadtgebietes zu den Naturräumen Geest und Marsch wird dargestellt',
             dataset.get('notes'))
-
+        # Publisher and ContactPoint
+        extras = dataset.get('extras')
+        self.assertTrue(len(extras) > 0)
+        self.assertIn(u'Behörde für Umwelt und Energie (BUE), Amt für Umweltschutz', self._get_value_from_extras(extras, 'publisher_name'))
+        self.assertIn(u'Herr Dr. Michael Schröder', self._get_value_from_extras(extras, 'contact_name'))
+        # Resources
         resources = dataset.get('resources')
         self.assertEqual(len(resources), 2)
         for res in resources:
@@ -619,3 +777,49 @@ class TestDCATde(unittest.TestCase):
             self.assertIn(
                 u'Das ist eine deutsche Beschreibung der Distribution',
                 res.get('description'))
+
+    def test_parse_dataset_dct_format_iana_uri(self):
+        resources = self._build_and_parse_format_mediatype_graph(
+            URIRef("https://www.iana.org/assignments/media-types/application/json")
+        )
+        # IANA mediatype URI should be added to mimetype field as well
+        self.assertIn(u'json', resources[0].get('format').lower())
+        self.assertEqual(u'https://www.iana.org/assignments/media-types/application/json',
+                         resources[0].get('mimetype'))
+
+    def test_parse_dataset_mediatype_iana_uri_without_format(self):
+        resources = self._build_and_parse_format_mediatype_graph(
+            mediatype_item=URIRef("https://www.iana.org/assignments/media-types/application/json")
+        )
+        # IANA mediatype URI should be added to mimetype field and to format as well
+        self.assertEqual(u'https://www.iana.org/assignments/media-types/application/json',
+                         resources[0].get('mimetype'))
+        self.assertEqual(u'https://www.iana.org/assignments/media-types/application/json',
+                         resources[0].get('format'))
+
+    def test_parse_dataset_dct_format_other_uri(self):
+        resources = self._build_and_parse_format_mediatype_graph(
+            URIRef("https://example.com/my/format")
+        )
+        self.assertEqual(u'https://example.com/my/format',
+                         resources[0].get('format'))
+        self.assertEqual(None, resources[0].get('mimetype'))
+
+    def test_parse_dataset_dct_format_mediatype_text(self):
+        resources = self._build_and_parse_format_mediatype_graph(
+            Literal("application/json")
+        )
+        # IANA mediatype should be added to mimetype field as well
+        self.assertIn(u'json', resources[0].get('format').lower())
+        self.assertEqual(u'application/json',
+                         resources[0].get('mimetype'))
+
+    def test_parse_dataset_format_and_dcat_mediatype(self):
+        # Even if dct:format is a valid IANA type, prefer dcat:mediaType if given
+        resources = self._build_and_parse_format_mediatype_graph(
+            Literal("application/json"), Literal("test-mediatype")
+        )
+        # both should be stored
+        self.assertIn(u'json', resources[0].get('format').lower())
+        self.assertEqual(u'test-mediatype',
+                         resources[0].get('mimetype'))